+++ samples/client/petstore/java/resteasy/src/test/java/io/swagger/client/api/PetApiTest.java
@@ -65,7 +65,7 @@ public class PetApiTest {
     @Test
     public void deletePetTest() throws ApiException {
         
-        Integer petId = null;
+        Long petId = null;
         
         String apiKey = null;
         
@@ -120,8 +120,8 @@ public class PetApiTest {
      */
     @Test
     public void getPetByIdTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         Pet response = api.getPetById(petId);
 
@@ -156,8 +156,8 @@ public class PetApiTest {
      */
     @Test
     public void updatePetWithFormTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         String name = null;
         String status = null;
         
@@ -176,8 +176,8 @@ public class PetApiTest {
      */
     @Test
     public void uploadFileTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         String additionalMetadata = null;
         java.io.File file = null;
